Refactor utils for later.
This will be used for handling errors during playback.
