+++ rust/src/utils/mod.rs
@@ -17,4 +17,4 @@
 pub mod playback_state_stream;
 pub mod progress_state_stream;
 pub mod metadata_callback_stream;
-pub mod playback_state;
+pub mod types;
